Refine pointToLayer call to include base feature used.
This fixes an issue where standar geometry had access to the base feature
through the geojson object, but geometry collections did not have access
to the feature that we are creating geometry for.

This could cause problems in user code that wants to use properties
from the feature to customize the Marker that is created for point
objects.